+++ src/routes/+error.svelte
@@ -1,12 +1,14 @@
 <script context="module">
-  /** @type {import('@sveltejs/kit').Load} */
-  export function load({ error, status }) {
-    return {
-      props: {
-        title: `${status}: ${String(error?.message)}`
-      }
-    };
-  }
+  throw new Error("@migration task: Replace error load function (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3293209)");
+
+  // /** @type {import('@sveltejs/kit').Load} */
+  // export function load({ error, status }) {
+  //   return {
+  //     props: {
+  //       title: `${status}: ${String(error?.message)}`
+  //     }
+  //   };
+  // }
 </script>
 
 <script>
