+++ src/routes/manifesto/+page.server.ts
@@ -1,3 +1,5 @@
+throw new Error("@migration task: Update +page.server.js (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292699)");
+
 import type { RequestHandler } from '@sveltejs/kit';
 
 // Redirect to wiki. Remove this file to bring back the original page.
