+++ src/routes/manifesto/+page.svelte
